requirements.txt
celery==5.3.1
redis==4.6.0
mysql-connector-python==8.1.0
#Instalation Redis and run it.

